requirements.txt
fastapi==0.95.0
uvicorn==0.23.1
mysql-connector-python==8.0.31
pydantic==1.10.7
